Rename module `slash` to `slashcommands`
