+++ features/messages/impl/build.gradle.kts
@@ -53,7 +53,7 @@ dependencies {
     implementation(projects.libraries.preferences.api)
     implementation(projects.libraries.recentemojis.api)
     implementation(projects.libraries.roomselect.api)
-    implementation(projects.libraries.slash.api)
+    implementation(projects.libraries.slashcommands.api)
     implementation(projects.libraries.audio.api)
     implementation(projects.libraries.voiceplayer.api)
     implementation(projects.libraries.voicerecorder.api)
@@ -105,5 +105,5 @@ dependencies {
     testImplementation(projects.features.poll.test)
     testImplementation(projects.libraries.eventformatter.test)
     testImplementation(projects.libraries.recentemojis.test)
-    testImplementation(projects.libraries.slash.test)
+    testImplementation(projects.libraries.slashcommands.test)
 }
