--- libraries/slash/api/build.gradle.kts
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) 2026 Element Creations Ltd.
- *
- * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
- * Please see LICENSE files in the repository root for full details.
- */
-plugins {
-    id("io.element.android-library")
-}
-
-android {
-    namespace = "io.element.android.libraries.slash.api"
-}
-
-dependencies {
-    implementation(projects.libraries.matrix.api)
-}
