--- libraries/slash/api/src/main/kotlin/io/element/android/libraries/slash/api/ChatEffect.kt
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2026 Element Creations Ltd.
- *
- * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
- * Please see LICENSE files in the repository root for full details.
- */
-
-package io.element.android.libraries.slash.api
-
-enum class ChatEffect {
-    CONFETTI,
-    SNOWFALL
-}
