--- libraries/slash/api/src/main/kotlin/io/element/android/libraries/slash/api/SlashCommandSuggestion.kt
@@ -1,14 +1,0 @@
-/*
- * Copyright (c) 2026 Element Creations Ltd.
- *
- * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
- * Please see LICENSE files in the repository root for full details.
- */
-
-package io.element.android.libraries.slash.api
-
-data class SlashCommandSuggestion(
-    val command: String,
-    val parameters: String?,
-    val description: String,
-)
