--- libraries/slash/test/build.gradle.kts
@@ -1,19 +1,0 @@
-/*
- * Copyright (c) 2026 Element Creations Ltd.
- *
- * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
- * Please see LICENSE files in the repository root for full details.
- */
-plugins {
-    id("io.element.android-library")
-}
-
-android {
-    namespace = "io.element.android.libraries.slash.test"
-}
-
-dependencies {
-    implementation(projects.libraries.slash.api)
-    implementation(projects.libraries.matrix.api)
-    implementation(projects.tests.testutils)
-}
