+++ libraries/slashcommands/impl/build.gradle.kts
@@ -1,0 +1,35 @@
+import extension.setupDependencyInjection
+import extension.testCommonDependencies
+
+/*
+ * Copyright (c) 2026 Element Creations Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
+ * Please see LICENSE files in the repository root for full details.
+ */
+plugins {
+    id("io.element.android-library")
+}
+
+android {
+    namespace = "io.element.android.libraries.slashcommands.impl"
+}
+
+setupDependencyInjection()
+
+dependencies {
+    implementation(projects.libraries.androidutils)
+    implementation(projects.libraries.preferences.api)
+    implementation(projects.libraries.core)
+    implementation(projects.libraries.matrix.api)
+    api(projects.libraries.slashcommands.api)
+    implementation(projects.libraries.di)
+    implementation(projects.libraries.featureflag.api)
+    implementation(projects.services.toolbox.api)
+
+    testCommonDependencies(libs)
+    testImplementation(projects.libraries.featureflag.test)
+    testImplementation(projects.libraries.preferences.test)
+    testImplementation(projects.libraries.matrix.test)
+    testImplementation(projects.services.toolbox.test)
+}
