+++ libraries/slashcommands/impl/src/main/res/values/temporary.xml
