+++ libraries/textcomposer/impl/src/main/kotlin/io/element/android/libraries/textcomposer/mentions/ResolvedSuggestion.kt
@@ -14,7 +14,7 @@ import io.element.android.libraries.designsystem.components.avatar.AvatarSize
 import io.element.android.libraries.matrix.api.core.RoomAlias
 import io.element.android.libraries.matrix.api.core.RoomId
 import io.element.android.libraries.matrix.api.room.RoomMember
-import io.element.android.libraries.slash.api.SlashCommandSuggestion
+import io.element.android.libraries.slashcommands.api.SlashCommandSuggestion
 
 @Immutable
 sealed interface ResolvedSuggestion {
